use internal endpoint for service to service communication (bsc#1116408)
So for Glance image signing, the rat hole goes like this

glance -> cursive -> castellan -> barbican

In other words, if you search the keyword barbican in the glance code base,
you won't find it. Instead, for barbican related options, we'll need to
look at castellan.

For barbican, the barbican_endpoint_type option in the [barbican]
of the service configuration file dictates the barbican endpoint to use.
By default, it is set to public. We'll need to set it to internal
instead as controllers have no access to the public network.

Change-Id: Id45838bdcd0f986c4da9483bcab2c644309e3fd6