SCRD-5578 Remove deprecated glance-registry
Issues: SCRD-5578

Remove support for deploying the glance-registry service which was
deprecated for removal in the OpenStack Queens release of Glance.

Doing so involves moving the setup of the database and metadefs from
the GLA-REG to the GLA-API role. Additionally the database connection
now needs to be setup in the glance-api.conf.j2 file.

Updated top-level playbooks to remove the GLA-REG related actions.

Updated the embedded set of metadefs files with the versions from the
OpenStack Rocky release of Glance.

Remove references to glance-registry from project stack (ardana-ci/)
input models.

Updated the config/glance-symlinks.yml to remove references to the
glance-registry related entries, and add/update links for all of the
metadefs json files.

NOTES:
  * Upgrading from Cloud8 to Cloud9 will require special handling
    to stop/disable the existing glance-registry service instances.
  * We can potentially remove the embedded metadefs files from the
    glance-ansible repo, and instead load the standard upstream ones.
    If we still want to support customising the metadefs, that could
    be achieved by additionally loading any customer supplied ones,
    stored under an appropriate path in the ansible code base, after
    the standard ones.

Change-Id: I92ff319189b162dddd757c1b3cda5f8a57fcacaa