Configure glance image_direct_url/multiple_locations (SOC-9285)
Although setting those configurations to true impose a security risk,
for environments where security is not a concern (virtual
deployments/CI) this option can be set to True to reduce the instance
boot time when ceph is used as backend.

Setting those optioins to True enable RBD cloning when creating
ephemeral disks for instances, in that way as the image does not need to be
downloaded to each compute and sent back to the ceph cluster when creating
an instance, the instance build time is drastically reduced.

Change-Id: I217402c6b0737e8a7e5aa058efe7ea170d8c1f22