Idempotent cirros image upload to glance (SOC-11342)
Every time the glance-cloud-configure.yml playbook is run it will
re-upload the cirros image to Glance, even if it already exists.
This can result in problems for any tools that expect to find only
a single cirros image such as the iverify.sh test.

Also move the cirros image name into a fact so that it is easier
to maintain, if the need arises.

Change-Id: I01ef9013dcf04908408956503e08f341847ebed8